Tree-shake string
