+++ lib/string.js
@@ -1,6 +1,10 @@
 'use strict';
 
-var util = require('./utils/handlebarsUtils');
+var util = {
+  isString: require('./utils/isString'),
+  isObject: require('./utils/isObject'),
+  options: require('./utils/options')
+};
 var utils = require('./utils');
 var helpers = module.exports;
