+++ lib/utils/handlebarsUtils.js
@@ -211,39 +211,7 @@ utils.isApp = function(thisArg) {
     && utils.isObject(thisArg.app);
 };
 
-/**
- * Creates an options object from the `context`, `locals` and `options.`
- * Handlebars' `options.hash` is merged onto the options, and if the context
- * is created by [templates][], `this.options` will be merged onto the
- * options as well.
- *
- * @param {Object} `context`
- * @param {Object} `locals` Options or locals
- * @param {Object} `options`
- * @return {Boolean}
- * @api public
- */
-
-utils.options = function(thisArg, locals, options) {
-  if (utils.isOptions(thisArg)) {
-    return utils.options({}, locals, thisArg);
-  }
-  if (utils.isOptions(locals)) {
-    return utils.options(thisArg, options, locals);
-  }
-  options = options || {};
-  if (!utils.isOptions(options)) {
-    locals = Object.assign({}, locals, options);
-  }
-  var opts = Object.assign({}, locals, options.hash);
-  if (utils.isObject(thisArg)) {
-    opts = Object.assign({}, thisArg.options, opts);
-  }
-  if (opts[options.name]) {
-    opts = Object.assign({}, opts[options.name], opts);
-  }
-  return opts;
-};
+utils.options = require('./options');
 
 /**
  * Get the context to use for rendering.
